+++ home-helper-api/src/main/kotlin/home/helper/api/persistence/database/mysql/table/DbHelpPointEarnedDetailMapper.kt
@@ -3,7 +3,14 @@
  */
 package home.helper.api.persistence.database.mysql.table
 
-import org.apache.ibatis.annotations.*
+import org.apache.ibatis.annotations.DeleteProvider
+import org.apache.ibatis.annotations.InsertProvider
+import org.apache.ibatis.annotations.Mapper
+import org.apache.ibatis.annotations.Result
+import org.apache.ibatis.annotations.ResultMap
+import org.apache.ibatis.annotations.Results
+import org.apache.ibatis.annotations.SelectProvider
+import org.apache.ibatis.annotations.UpdateProvider
 import org.apache.ibatis.type.JdbcType
 import org.mybatis.dynamic.sql.delete.render.DeleteStatementProvider
 import org.mybatis.dynamic.sql.insert.render.InsertStatementProvider
@@ -34,8 +41,8 @@ interface DbHelpPointEarnedDetailMapper {
     @Results(
         id = "DbHelpPointEarnedDetailRecordResult", value = [
             Result(column = "id", property = "id", jdbcType = JdbcType.BIGINT, id = true),
-            Result(column = "achievement_id", property = "achievementId", jdbcType = JdbcType.INTEGER),
-            Result(column = "item_id", property = "itemId", jdbcType = JdbcType.INTEGER)
+            Result(column = "achievement_id", property = "achievementId", jdbcType = JdbcType.BIGINT),
+            Result(column = "item_id", property = "itemId", jdbcType = JdbcType.BIGINT)
         ]
     )
     fun selectMany(selectStatement: SelectStatementProvider): List<DbHelpPointEarnedDetailRecord>
